requirements.txt
discord.py~=2.5.2
certifi~=2025.1.31
PyPDF2~=1.26.0
requests~=2.32.0
pathlib~=1.0.1
PyMuPDF~=1.25.3
urllib3~=2.2.3